Improve model naming
