+++ src/Model/Tweet/TweetUpdated.php
@@ -12,43 +12,32 @@ use FAPI\Boilerplate\Model\CreatableFromArray;
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  */
-class IndexResponse implements CreatableFromArray
+class TweetUpdated implements CreatableFromArray
 {
     private $message;
 
-    private $tweets;
-
     /**
      * @param string $message
-     * @param array  $tweets
      */
-    private function __construct(string $message, array $tweets)
+    private function __construct(string $message)
     {
         $this->message = $message;
-        $this->tweets = $tweets;
     }
 
     /**
      * @param array $data
      *
-     * @return IndexResponse
+     * @return TweetUpdated
      */
     public static function createFromArray(array $data)
     {
         $message = '';
-        $tweets = [];
-
-        if (isset($data['tweets'])) {
-            foreach ($data['tweets'] as $item) {
-                $tweets[] = Tweet::create($item);
-            }
-        }
 
         if (isset($data['message'])) {
             $message = $data['message'];
         }
 
-        return new self($message, $tweets);
+        return new self($message);
     }
 
     /**
@@ -57,13 +46,5 @@ class IndexResponse implements CreatableFromArray
     public function getMessage(): string
     {
         return $this->message;
-    }
-
-    /**
-     * @return Tweet[]
-     */
-    public function getTweets(): array
-    {
-        return $this->tweets;
     }
 }
