+++ src/Model/Tweet/Tweets.php
@@ -12,32 +12,43 @@ use FAPI\Boilerplate\Model\CreatableFromArray;
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  */
-class CreateResponse implements CreatableFromArray
+class Tweets implements CreatableFromArray
 {
     private $message;
 
+    private $tweets;
+
     /**
      * @param string $message
+     * @param array  $tweets
      */
-    private function __construct(string $message)
+    private function __construct(string $message, array $tweets)
     {
         $this->message = $message;
+        $this->tweets = $tweets;
     }
 
     /**
      * @param array $data
      *
-     * @return CreateResponse
+     * @return Tweets
      */
     public static function createFromArray(array $data)
     {
         $message = '';
+        $tweets = [];
+
+        if (isset($data['tweets'])) {
+            foreach ($data['tweets'] as $item) {
+                $tweets[] = Tweet::create($item);
+            }
+        }
 
         if (isset($data['message'])) {
             $message = $data['message'];
         }
 
-        return new self($message);
+        return new self($message, $tweets);
     }
 
     /**
@@ -46,5 +57,13 @@ class CreateResponse implements CreatableFromArray
     public function getMessage(): string
     {
         return $this->message;
+    }
+
+    /**
+     * @return Tweet[]
+     */
+    public function getTweets(): array
+    {
+        return $this->tweets;
     }
 }
